Documentation

This is the documentation for StellaratorNearAxis.jl structs and their operations.

Contents

Documentation

Contents

Index

Structs

Constructors

Operations

Helper Functions

Coil Functions

Direct Frenet-Serret Equilibrium Problem

Plotting

Index

StellaratorNearAxis.Coil

StellaratorNearAxis.DirectNearAxisEquilibrium

StellaratorNearAxis.FluxPowerSeries

StellaratorNearAxis.IdentityPowerSeries

StellaratorNearAxis.PowerSeriesRho

StellaratorNearAxis.SpatialPowerSeries

StellaratorNearAxis.SpatialPowerSeries

StellaratorNearAxis.SpectralPowerSeries

StellaratorNearAxis.SpectralPowerSeries

StellaratorNearAxis.ZeroPowerSeries

Base.:*

Base.:*

Base.:+

Base.:-

Base.:-

Base.:/

Base.:^

Base.Math.sincos

Base.cos

Base.cosh

Base.div

Base.exp

Base.inv

Base.similar

Base.sin

Base.sinh

LinearAlgebra.norm

StellaratorNearAxis.Fnorm

StellaratorNearAxis.InitialVacuumNearAxisEquilibrium

StellaratorNearAxis.L2norm

StellaratorNearAxis.change_order

StellaratorNearAxis.compose

StellaratorNearAxis.composition_basis

StellaratorNearAxis.distribute_p0

StellaratorNearAxis.evaluate

StellaratorNearAxis.evaluate

StellaratorNearAxis.field_to_nae

StellaratorNearAxis.find_magnetic_axis

StellaratorNearAxis.fit_SpectralPowerSeries

StellaratorNearAxis.flux_compose

StellaratorNearAxis.flux_contours!

StellaratorNearAxis.flux_surface_plot!

StellaratorNearAxis.get_field_on_axis

StellaratorNearAxis.get_flux_coordinates

StellaratorNearAxis.grad

StellaratorNearAxis.invert_coordinates

StellaratorNearAxis.load_coils

StellaratorNearAxis.magnetic_trajectory

StellaratorNearAxis.plot_curve!

StellaratorNearAxis.remove_zeros

StellaratorNearAxis.rho_antideriv

StellaratorNearAxis.rho_deriv

StellaratorNearAxis.rho_integrate

StellaratorNearAxis.s_deriv

StellaratorNearAxis.surface_integrate

StellaratorNearAxis.theta_deriv

StellaratorNearAxis.unsafe_distribute_p0

StellaratorNearAxis.vacuum_solve

StellaratorNearAxis.volume_integrate

StellaratorNearAxis.zero_FluxPowerSeries

StellaratorNearAxis.zero_SpatialPowerSeries

StellaratorNearAxis.zero_SpectralPowerSeries

Structs

StellaratorNearAxis.PowerSeriesRho — Type

PowerSeriesRho

A struct used for algebraically representing changes in order of power series, typically stored as p0.

Example

A = zero_SpectralPowerSeries(1,1)
A[1].a[1] = 1.         # A is a constant SpectralPowerSeries
display(get_p0(A))     # The leading order of A is 0
ρ = PowerSeriesRho()
B = ρ^2*A
display(get_p0(B))     # The SpectralPowerSeries B is ρ^2

StellaratorNearAxis.ZeroPowerSeries — Type

ZeroPowerSeries{T} <: AbstractPowerSeries{T}

A power series representing the zero element. Multiplication of an AbstractPowerSeries by a ZeroPowerSeries returns a ZeroPowerSeries, and addition of an AbstractPowerSeries A with a ZeroPowerSeries returns A.

Example

A = zero_SpectralPowerSeries(1,1)
A[1].a[1] = 1.
Z = ZeroPowerSeries()
display(A*Z) # Returns a ZeroPowerSeries
display(A+Z) # Returns A

StellaratorNearAxis.IdentityPowerSeries — Type

IdentityPowerSeries{T} <: AbstractPowerSeries{T}

A power series representing the identity element. Multiplication of an AbstractPowerSeries A by an IdentityPowerSeries returns A.

Example

A = zero_SpectralPowerSeries(1,1)
A[1].a[1] = 2.
id = IdentityPowerSeries()
display(A*id) # Returns A

StellaratorNearAxis.FluxPowerSeries — Type

FluxPowerSeries{T} <: AbstractPowerSeries{T}

Formal power series that is a function of the flux surface, i.e. A = A₀ + A₁ (ρ²)¹ + A₂ (ρ²)² + ...

StellaratorNearAxis.SpectralPowerSeries — Type

SpectralPowerSeries{T} <: AbstractPowerSeries{T}

The basic spectral representation of near-axis series. A SpectralPowerSeries is composed of SpectralPowerSeriesSlices, each of which can be accessed by getindex via A[n]. Each slice n contains an M × n array of Fourier coefficients, where M is the number of coefficients in the s direction and n is the number of coefficients in the θ direction. In the s direction, the coefficients are ordered as 1, sin(s), cos(s), sin(2s), ... In the odd n case in the θ direction, the coefficients are ordered as 1, sin(2θ), cos(2θ), sin(4θ), ... In the even n case in the θ direction, the coefficients are ordered as cos(θ), sin(θ), cos(3θ), sin(3θ), ... Derivative routines are defined for SpectralPowerSeries; see rho_deriv, theta_deriv, s_deriv, grad, div.

Example

Create a power series representing y = ρ sin(θ)

M = 5                              # Number of modes in s
N = 3                              # Number of orders in rho
y = zero_SpectralPowerSeries(M, N) # Create a SpectralPowerSeries
y[2].a[1,2] = 1.                   # Set the ρ sin(θ) coefficient to 1
y

StellaratorNearAxis.SpatialPowerSeries — Type

SpatialPowerSeries{T} <: AbstractPowerSeries{T}

The representation of near-axis series on collocation nodes. A SpatialPowerSeries is composed of SpatialPowerSeriesSlices, each of which can be accessed by getindex via A[n]. Each slice n contains an M × n array of sampled values, where M is the number of samples in the s direction and n is the order in the θ direction. The s points are ordered as 2(j-1)π/M for 1≤j≤M, while in the θ points are ordered as 2π(k-1)/(2n-1) for 1≤k≤n. Algebraic operations, such as *, /, inv, etc. are defined for SpatialPowerSeries

Example

Create a power series representing y = ρ sin(θ)

Mc = 5                             # Number of points in s
N = 3                              # Number of orders in rho
y = zero_SpectralPowerSeries(1, N) # Create a SpectralPowerSeries
y[2].a[1,2] = 1.                   # Set the ρ sin(θ) coefficient to 1
y_spatial = SpatialPowerSeries(y, M=Mc)
y_spatial

Constructors

StellaratorNearAxis.zero_SpectralPowerSeries — Function

zero_SpectralPowerSeries(M::Integer, N::Integer; p0::Integer = 0)

Return a SpectralPowerSeries object initialized with zero coefficients.

Inputs:

M: Number of Fourier coefficients in the poloidal s coordinate

N: Number of coefficients in radial ρ and toroidal θ coordinates

p0: The leading order ρ coefficient

Output:

A zero SpectralPowerSeries

StellaratorNearAxis.zero_SpatialPowerSeries — Function

zero_SpatialPowerSeries(M::Integer, N::Integer; p0::Integer = 0)

Return a SpatialPowerSeries object initialized with zero entries.

Inputs:

M: Number of nodes in the poloidal s coordinate

N: Order in radial ρ and toroidal θ coordinates

p0: The leading order ρ coefficient

Output:

A zero SpatialPowerSeries

StellaratorNearAxis.zero_FluxPowerSeries — Function

zero_FluxPowerSeries(N::Integer; p0::Integer=0)

Create a FluxPowerSeries of degree N and leading coefficient p0 with zero entries.

Base.similar — Method

similar(A<:AbstractPowerSeries; M=nothing, N=nothing, p0=nothing)

Create a zero AbstractPowerSeries of the same type as A (except, in the case of an IdentityPowerSeries, returns another IdentityPowerSeries). Can optionally set the number of Spatial/Spectral modes M, series order N, and order offset p0.

StellaratorNearAxis.SpatialPowerSeries — Method

SpatialPowerSeries(A::SpectralPowerSeries{T}; M::Integer = -1) where {T}

Transform a SpectralPowerSeries to a SpatialPowerSeries of the same order. By default, the output uses the same number of collocation nodes as the Fourier modes. Alternatively, the user can set M to choose a different number of nodes.

StellaratorNearAxis.SpectralPowerSeries — Method

SpectralPowerSeries(A::SpatialPowerSeries{T}; M::Integer=-1) where {T}

Transform a SpatialPowerSeries to a SpectralPowerSeries of the same order. By default, the output uses the same number of Fourier modes as the number of spatial collocation nodes. Alternatively, the user can set M to choose a different number of modes.

StellaratorNearAxis.fit_SpectralPowerSeries — Function

fit_SpectralPowerSeries(a::AbstractArray, r::AbstractVector, M::Integer, N::Integer)

Fit a SpectralPowerSeries to the samples of some function.

input:

a: A Nr × Nθ × Nϕ array of function samples. It is assumed that a[i,j,k] = f(r[i], θ[j], ϕ[k]), where r is given as input, and θ and ϕ are on a uniform grid from 0 to 2π

r: The radial coordinates sampled at

N: The order in ρ and θ the series is interpolated to

M: The order in ϕ the series is interpolated to

fit_SpectralPowerSeries(a::AbstractArray, r::AbstractVector, M::Integer, N::Integer)

Fit a SpectralPowerSeries to the samples of some function.

input:

f: A function from R^3→R, with signature f(r,θ,ϕ)

r: The radial coordinates sampled at

Nθ: Number of θ samples

Nϕ: Number of ϕ samples

N: The order in ρ and θ the series is interpolated to

M: The order in ϕ the series is interpolated to

Operations

StellaratorNearAxis.evaluate — Method

evaluate(A::SpectralPowerSeries, rho::Union{Number, AbstractVector},
         theta::Union{Number, AbstractVector}, phi::Union{Number, AbstractVector})

Evaluate the SpectralPowerSeries either at a point or on a grid.

Base.:- — Method

-(A::AbstractPowerSeries)

Get the negative power series of A (i.e., A + (-A) == 0).

Base.:+ — Method

+(A::AbstractPowerSeries, B::AbstractPowerSeries)

Add two AbstractPowerSeries. May not work for all inputs (e.g. a SpectralPowerSeries and a SpatialPowerSeries).

Base.:- — Method

-(A::AbstractPowerSeries, B::AbstractPowerSeries)

Subtract two AbstractPowerSeries. May not work for all inputs (e.g. a SpectralPowerSeries and a SpatialPowerSeries).

Base.:* — Method

*(A::AbstractPowerSeries{T}, B::AbstractPowerSeries{T};
       N::Integer=-1, careful::Bool=false, careful_tol::Number=1e-12) where {T}

Multiply two power series together to get C = A*B to order N. If careful==true, the output order is determined by the leading order by which A and B are nonzero to tolerance careful_tol. In this case, the argument for N is ignored.

Base.:* — Method

*(a::Number, A::AbstractPowerSeries)

Multiply the power series A by a scalar number.

Base.inv — Method

inv(A::AbstractPowerSeries; N::Integer = -1)

Find the multiplicative inverse of A to order N. Approximately satisfies A*inv(A)=1

Base.:/ — Method

/(A::AbstractPowerSeries{T}, B::AbstractPowerSeries{T};
       N::Integer=-1, careful::Bool=false, careful_tol::Number=1e-12) where {T}

Divide two power series to get C = A/B to order N. Currently just calls inv() and *(), but could potentially be faster with a better method.

Base.exp — Method

exp(A::AbstractPowerSeries{T}[, alpha::T]; N::Integer = -1) where {T}

Exponentiate the power series A to get e^(αA) to order N.

Base.sinh — Method

sinh(A::AbstractPowerSeries{T}[, alpha::T]) where {T}

Find sinh(αA) = (e^(αA) - e^(-αA))/2

Base.cosh — Method

cosh(A::AbstractPowerSeries{T}[, alpha::T]) where {T}

Find cosh(αA) = (e^(αA) + e^(-αA))/2.

Base.Math.sincos — Method

sincos(A::AbstractPowerSeries[, ω::Number])

Find both the sine and cosine of A, i.e. sin(ωA), cos(ωA) = sincos(A[, ω]) (note: both sin and cos call sincos), so calling this can save time if you require both)

Base.sin — Method

sin(A::AbstractPowerSeries[, ω::Number])

Finds the sine of a power series sin(ωA). Note: if you also require cos, it is twice as efficient to call sincos.

Base.cos — Method

cos(A::AbstractPowerSeries[, ω::Number])

Finds the cosine of a power series cos(ωA). Note: if you also require sin, it is twice as efficient to call sincos.

Base.:^ — Method

^(A::AbstractPowerSeries, alpha::Number; N::Integer=-1)

Raise A to the power α to N terms in the power series, i.e. compute A^α.

StellaratorNearAxis.rho_deriv — Function

rho_deriv(A::AbstractPowerSeries{T}) where {T}

Finds the derivative of A with respect to ρ.

StellaratorNearAxis.theta_deriv — Function

theta_deriv(A::SpectralPowerSeries{T}) where {T}

Takes the derivative of A in the θ direction.

StellaratorNearAxis.s_deriv — Function

s_deriv(A::SpectralPowerSeries{T}) where {T}

Takes the derivative of A with respect to s.

WARNING: currently this function is only defined for an odd number of Fourier modes.

StellaratorNearAxis.grad — Function

grad(A::SpectralPowerSeries{T}) where {T}

Returns the "gradient" of [dA/dρ, dA/dθ, dA/ds]. Note that for a true gradient, it is necessary to use the metric.

Base.div — Function

div(A::SpectralPowerSeries{T}) where {T}

Returns the "divergence" dA[1]/dρ + dA[2]/dθ + dA[3]/ds. Note that for a true divergence, it is necessary to use the metric.

StellaratorNearAxis.surface_integrate — Function

surface_integrate(A::AbstractPowerSeries)

Performs the surface integral ∫A dθ dϕ.

StellaratorNearAxis.rho_integrate — Function

rho_integrate(A::AbstractPowerSeries)

Perform the integral B(r) = ∫₀ʳ A(ρ) ρ dρ See rho_antideriv for the integral without the integrating factor.

StellaratorNearAxis.rho_antideriv — Method

rho_antideriv(A::AbstractPowerSeries)

Perform the integral B(r) = ∫₀ʳ A(ρ) dρ (note: unlike rho_integrate, which includes an integrating factor for analytic integrals)

StellaratorNearAxis.volume_integrate — Function

volume_integate(A::AbstractPowerSeries{T}; N::Integer=-1) where {T}

Integral over a volume. Implemented by composing rho_integrate with surface_integrate.

LinearAlgebra.norm — Method

LinearAlgebra.norm(A::AbstractVector{<:AbstractPowerSeries})

Returns the norm of a power series A, i.e. (A'*A)^0.5. For a norm of the elements of A, see Fnorm.

StellaratorNearAxis.Fnorm — Function

Fnorm(A::AbstractPowerSeries{T}) where {T}

Frobenius norm of the elements of A. Useful for determining approximate equality of AbstractPowerSeries.

StellaratorNearAxis.L2norm — Method

L2norm(A::AbstractPowerSeries, δ::Number)

Find the L2 norm of a power series via near-axis operations. NOTE: this is not the true L2 norm; just the near-axis approximation of it.

StellaratorNearAxis.flux_compose — Method

flux_compose(A::FluxPowerSeries{T}, B::SpatialPowerSeries{T}) where {T}

Composes a FluxPowerSeries A with a SpatialPowerSeries B, i.e. A∘B. Useful, e.g., for composing the rotational transform with the flux.

StellaratorNearAxis.composition_basis — Function

composition_basis(F::AbstractVector{<:SpectralPowerSeries})

Get a basis of the form -[[1], [ρF cos(θF), ρF sin(θF)], [ρF^2, ρF^2 cos(2 θF), ρF^2 sin(2 θF)], ...] from the input of a SpectralPowerSeries -F = [ρF cos(θF), ρF sin(θF)] This can be used in conjunction with the function compose to find the composition of another SpectralPowerSeries with F.

StellaratorNearAxis.compose — Function

compose(A::SpectralPowerSeries, Ftree::AbstractVector)

Compose a SpectralPowerSeries A with the basis Ftree. See composition_basis for a method to obtain Ftree from a change-of-coordinates SpectralPowerSeries F.

StellaratorNearAxis.invert_coordinates — Function

invert_coordinates(F::AbstractVector)

Invert the SpectralPowerSeries coordinate transform F. That is, the composition via compose of F with the output G satisfies F∘G(x) ≈ x

Helper Functions

StellaratorNearAxis.remove_zeros — Method

remove_zeros(A::AbstractPowerSeries, tol=1e-10)

If one has access to a power series A = ρ^2N B, returns the power series B. For example, when the polar Laplacian is applied to a series, the output has zeros in the first two orders. This removes those zeros.

StellaratorNearAxis.change_order — Method

change_order(A::AbstractPowerSeries{T}, N::Integer) where {T}

Creates a new PowerSeries with the coefficients of A with order N. If N > get_N(A), the extra coefficients are set to zero.

StellaratorNearAxis.distribute_p0 — Method

distribute_p0(A::AbstractPowerSeries, p0::Integer)

If A has an order q0 ≥ p0, creates a new AbstractPowerSeries of the same type with order offset p0with equal coefficients. For analyticity to be preserved, we additionally require thatq0andp0are even offsets of each other (p0-q0 = 0 mod 2). Ifp0is not specified, the default value isp0=0`.

StellaratorNearAxis.unsafe_distribute_p0 — Function

unsafe_distribute_p0(A::AbstractPowerSeries[, p0::Integer])

Creates a new PowerSeries with order offset p0 with equal coefficients to A. For analyticity to be preserved, we require that q0 and p0 are even offsets of each other (p0-q0 = 0 mod 2). If p0 is not specified, the default value is p0=0.

If you are unsure about whether p0 can be increased without losing information, remove_zeros can automatically check whether the order can be changed.

Coil Functions

StellaratorNearAxis.Coil — Type

Coil

A coil consists of a filament and a current. It can be constructed by Coil(r0, J, Mcoil), where r0 is a 3-vector of SpectralPowerSeries, J is a number expressing the current per unit length, and Mcoil is the number of quadrature nodes for Biot-Savart.

StellaratorNearAxis.load_coils — Function

Input:

coil_file: A JSON file holding the output of a Simsopt coil optimization

Mcoil: The number of coil quadrature nodes for Biot-Savart

Output:

The coil set contained in coil_file, which can be used to initialize a near axis expansion.

StellaratorNearAxis.evaluate — Method

evaluate(x::AbstractArray, c::Coil)

Get the magnetic field from the coil c at points x in a 3×N array. Uses Biot-Savart discretized via the trapezoidal rule.

StellaratorNearAxis.magnetic_trajectory — Function

magnetic_trajectory(cs::Vector{Coil}, x0::AbstractVector, S::Number; tol::Number=1e-8)

Evolve the magnetic field generated by the coil set cs starting at the point x0 a distance S to tolerance tol. Outputs a OrdinaryDiffEq solution object. It also returns the times at which the trajectory passes through the Poincare surface intersects with the initial toroidal coordinate.

StellaratorNearAxis.find_magnetic_axis — Function

find_magnetic_axis(x0::AbstractVector, cs::Vector{Coil})

Find a point on the magnetic axis of a coil set cs with Mcoil quadrature points. Requires an initial guess of a point on the axis x0.

StellaratorNearAxis.get_field_on_axis — Function

get_field_on_axis(r0::AbstractVector, cs::Vector{Coil}, Mc::Number, N::Number, Mcoil::Number)

Expand the magnetic field from the coil set cs about the axis r0 (found, e.g., via find_magnetic_axis). The field is computed on Mc collocation points and converted to a SpectralPowerSeries of order Ms dictated by the value of Ms used to define r0. The field is computed to order N and with Mcoil quadrature points on the coils.

Output:

B: The magnetic field in Frenet-Serret coordinates

r: The Frenet-Serret coordinates

StellaratorNearAxis.field_to_nae — Function

field_to_nae(r0_s::AbstractVector{SpectralPowerSeries{T}}, 
    Bsup_s::AbstractVector{SpectralPowerSeries{T}}, Mc::Integer, K_reg::Integer, 
    N_reg::Integer) where {T}

Create a DirectNearAxisEquilibrium object with the magnetic field from a coil expansion.

Direct Frenet-Serret Equilibrium Problem

StellaratorNearAxis.DirectNearAxisEquilibrium — Type

DirectNearAxisEquilibrium

The struct used to represent an equilibrium of the direct near-axis expansion in vacuum. See InitialVacuumNearAxisEquilibrium to initialize the equilibrium. See DirectFS.jl for member descriptions. We use the convention that _s refers to SpectralPowerSeries and _c for SpatialPowerSeries (c for collocation).

StellaratorNearAxis.InitialVacuumNearAxisEquilibrium — Function

InitialVacuumNearAxisEquilibrium(r0_s::Vector{SpectralPowerSeries{T}},
           Nρ::Integer, Mc::Integer, B0_s::SpectralPowerSeries{T}, 
           K_reg::Number, N_reg::Integer) where {T}

Initialize a DirectNearAxisEquilibrium, which can then be solved via vacuum_solve. Input:

r0_s: The axis r0. The number of spectral modes determines the number of modes Ms for other SpectralPowerSeries objects.

Nρ: The order of the series

Mc: The number of collocation nodes for SpatialPowerSeries in the pseudospectral method

B0_s: The on-axis magnetic field magnitude

K_reg: The regularization wavenumber cutoff (set to Inf for no regularization)

N_reg: The regularization exponent

StellaratorNearAxis.vacuum_solve — Function

vacuum_solve(nae::DirectNearAxisEquilibrium, phi_IC::AbstractArray)

Solve for a vacuum field. If Ms is the number of Fourier modes, and Nρ is the order of the series the boundary conditions at each order are given by the Ms × 2Nρ-1 matrix

phi_IC = [phi_0c0, phi_1c1, phi_1s1, phi_2c2, phi_2s2, ..., phi_NcN, phi_NsN]

Note that the magnetic axis to truely be a magnetic axis, the first three columns are zero.

StellaratorNearAxis.get_flux_coordinates — Function

get_flux_coordinates(nae::DirectNearAxisEquilibrium)

Given a solved DirectNearAxisEquilibrium, find flux coordinates. They can be accessed via nae.ξ_s, and the flux/rotational transform can be found via nae.psi_s/nae.iota.

Plotting

StellaratorNearAxis.plot_curve! — Function

plot_curve!(ax, r0::Vector, M::Number; color=:black)

Plot a 3D curve from a vector of SpectralPowerSeries r0. Uses M mesh points. Can be used for both near-axis equilibria (r0 = nae.r0_s) or coils (r0 = coil.r0_s).

StellaratorNearAxis.flux_surface_plot! — Function

flux_surface_plot!(ax, nae::DirectNearAxisEquilibrium, u::SpectralPowerSeries, rho0::Number, 
                       theta0::Number, s0::Number; Ntheta::Integer=50, Ns::Integer=50, 
                       colormap=:diverging_isoluminant_cjo_70_c25_n256, kwargs...)

3D plot a flux surface of a near-axis equilibrium file. Input:

ax: A 3D CairoMakie CairoMakie.Axis3 struct

u: The SpectralPowerSeries to be plotted in direct coordinates (e.g. nae.phi_s)

(rho0,theta0,s0): A spatial point the flux surface (approximately) intersects

(Ntheta,Ns): The resolution of the plot

(colormap,kwargs): Additional arguments for CairoMakie.mesh!

StellaratorNearAxis.flux_contours! — Function

flux_contours!(ax, xi_s::AbstractVector, rho0::Number, theta0::Number, s0::Number, 
                    Nsurf::Integer; Ntheta = 101, color=:black, kwargs...)

Contour plot of flux surfaces on a cross section Input:

ax: A 3D CairoMakie CairoMakie.Axis struct

xi_s: The SpectralPowerSeries vector defining the flux surfaces (e.g. nae.xi_s)

(rho0,theta0): A point on a Poincare section which the outermost surface intersects

s0: The Poincare section taken

Nsurf: The number of surfaces to plot

Ntheta: Number of points used to plot each surface

(color,kwargs): Additional arguments for CairoMakie.lines!